requirements.txt
pika==1.3.1
python-dateutil==2.8.2
six==1.16.0
